Removed the "remove mutation before input"-optimization
Unit tests uncovered that the optimization did not actually optimize as
aggressively as it should have. The optimization removed at most one
mutation before input instead of removing all preceding mutation.
Instead of leaving the half-baked implementation in, it is removed with
the intention of re-adding it in the future.